distsqlrun: move common funcs from indexbackfiller
... into backfiller.go. Rename test file as well since it's only testing
the common functions.
