+++ pkg/sql/distsqlrun/indexbackfiller.go
@@ -27,7 +27,6 @@ import (
 	"github.com/cockroachdb/cockroach/pkg/sql/parser"
 	"github.com/cockroachdb/cockroach/pkg/sql/sqlbase"
 	"github.com/cockroachdb/cockroach/pkg/util/log"
-	"github.com/cockroachdb/cockroach/pkg/util/protoutil"
 	"github.com/cockroachdb/cockroach/pkg/util/timeutil"
 	"github.com/cockroachdb/cockroach/pkg/util/tracing"
 )
@@ -102,25 +101,6 @@ func (ib *indexBackfiller) nextRow(ctx context.Context) (sqlbase.EncDatumRow, er
 		return nil, err
 	}
 	return fetcherRow, nil
-}
-
-// ConvertBackfillError returns a cleaner SQL error for a failed Batch.
-func ConvertBackfillError(tableDesc *sqlbase.TableDescriptor, b *client.Batch) error {
-	// A backfill on a new schema element has failed and the batch contains
-	// information useful in printing a sensible error. However
-	// ConvertBatchError() will only work correctly if the schema elements
-	// are "live" in the tableDesc.
-	desc := protoutil.Clone(tableDesc).(*sqlbase.TableDescriptor)
-	mutationID := desc.Mutations[0].MutationID
-	for _, mutation := range desc.Mutations {
-		if mutation.MutationID != mutationID {
-			// Mutations are applied in a FIFO order. Only apply the first set
-			// of mutations if they have the mutation ID we're looking for.
-			break
-		}
-		desc.MakeMutationComplete(mutation)
-	}
-	return sqlbase.ConvertBatchError(desc, b)
 }
 
 // runChunk returns the next-key and an error. next-key is nil
@@ -273,79 +253,4 @@ func (ib *indexBackfiller) Run(ctx context.Context, wg *sync.WaitGroup) {
 		ib.output.Push(nil /* row */, ProducerMetadata{Err: err})
 	}
 	ib.output.ProducerDone()
-}
-
-// WriteResumeSpan writes a checkpoint for the backfill work on origSpan.
-// origSpan is the span of keys that were assigned to be backfilled,
-// resume is the left over work from origSpan.
-func WriteResumeSpan(
-	ctx context.Context,
-	db *client.DB,
-	id sqlbase.ID,
-	origSpan roachpb.Span,
-	resume roachpb.Span,
-	mutationIdx int,
-) error {
-	ctx, traceSpan := tracing.ChildSpan(ctx, "checkpoint")
-	defer tracing.FinishSpan(traceSpan)
-	if resume.Key != nil && !resume.EndKey.Equal(origSpan.EndKey) {
-		panic("resume must end on the same key as origSpan")
-	}
-	cnt := 0
-	return db.Txn(ctx, func(ctx context.Context, txn *client.Txn) error {
-		cnt++
-		tableDesc, err := sqlbase.GetTableDescFromID(ctx, txn, id)
-		if err != nil {
-			return err
-		}
-		if cnt > 1 {
-			log.Infof(ctx, "retrying adding checkpoint %s to table %s", resume, tableDesc.Name)
-		}
-
-		// This loop is finding a span in the checkpoint that fits
-		// origSpan. It then carves a spot for origSpan in the
-		// checkpoint, and replaces origSpan in the checkpoint with
-		// resume.
-		mutation := &tableDesc.Mutations[mutationIdx]
-		for i, sp := range mutation.ResumeSpans {
-			if sp.Key.Compare(origSpan.Key) <= 0 &&
-				sp.EndKey.Compare(origSpan.EndKey) >= 0 {
-				// origSpan is in sp; split sp if needed to accommodate
-				// origSpan and replace origSpan with resume.
-				before := mutation.ResumeSpans[:i]
-				after := append([]roachpb.Span{}, mutation.ResumeSpans[i+1:]...)
-
-				// add span to before, but merge it with the last span
-				// if possible.
-				addSpan := func(begin, end roachpb.Key) {
-					if begin.Equal(end) {
-						return
-					}
-					if len(before) > 0 && before[len(before)-1].EndKey.Equal(begin) {
-						before[len(before)-1].EndKey = end
-					} else {
-						before = append(before, roachpb.Span{Key: begin, EndKey: end})
-					}
-				}
-
-				// The work done = [origSpan.Key...resume.Key]
-				addSpan(sp.Key, origSpan.Key)
-				if resume.Key != nil {
-					addSpan(resume.Key, resume.EndKey)
-				} else {
-					log.VEventf(ctx, 2, "completed processing of span: %+v", origSpan)
-				}
-				addSpan(origSpan.EndKey, sp.EndKey)
-				mutation.ResumeSpans = append(before, after...)
-
-				log.VEventf(ctx, 2, "ckpt %+v", mutation.ResumeSpans)
-				txn.SetSystemConfigTrigger()
-				return txn.Put(ctx, sqlbase.MakeDescMetadataKey(tableDesc.GetID()), sqlbase.WrapDescriptor(tableDesc))
-			}
-		}
-		// Unable to find a span containing origSpan?
-		return errors.Errorf(
-			"span %+v not found among %+v", origSpan, mutation.ResumeSpans,
-		)
-	})
 }
